Better character state method
track character's current state with local attributes rather than command sets a previously proposed.
Idea loosely taken from https://github.com/evennia/evennia/wiki/Coding-FAQ#preventing-character-from-moving-based-on-a-condition
Should be a `state` object. That can be added as an attribute to a character object. It will hold the attributes like stunned busy dead so on. Those can then contain command keys that should be locked out if the state is true.