Interuptable command
created the outline of a method for commands that can be interupted.
The intention is to allow most commands to be interupted.
For example the user tries to teleport but gets punched in the face an fails a  concentration save so the command does not complete.
Another example would be a big slow attack prompts the target if they would like to spend fatigue to get a dodge bonus.