scripts, global script tags search
Created following
https://github.com/evennia/evennia/wiki/Scripts

Weather is an example of a object script.
DunWeather, (typo no need to fix for a learning mud) is an example of a global script.

Both echo a list of weather to the msg_contents of a room.
Weather by a randomly created list to the self.obj.msg_contents where obj is the automatically created attribute of the object the script is attached to.
DunWeather does it by collecting a list of rooms with the `dune_room` tag of category "location tags". Than looping through that list sending the message out. This later method would be an exmple of how you would actually do weather.

For now I had to add DunWeather to GOLBAL_SCRIPTS using a command that calls create_script. I will figure out how to add it directly to GLOBAL_SCRIPTS at a later time. The scripts tutorial was rather vague on that single step.