Command defferal, Character status
created system to defer Commands
creates system for Character status
in commands.command.Command the mux parse example method was uncommented. The parse was not occuring without it.
Started the creation of developer commands, mainly intended for unit testing. Good for reference also.
unit testing for Character status and Command defferal is in commands.tests.TestCommands