added mixins, redundant targetable, objects optional targetable
moved the targetable attribute to the database for persistence. Moved it to a mixin for Objects and Exits.
Created mixins ExitObjectAndRoomMixin and ExitAndObjectMixin to easily share among those object types
Added made objects being action targetable optional.
Moved usdesc for Objects, Rooms and Exits to the mixin ExitObjectAndRoomMixin, because the mixins are identical.