UM Clothing upgrade
Changed clothing type limits
created unit testing for clothing command testing in typeclass temporarily.
Clothing class inheirts typeclasses.objects.Object
Removed a strangely large number of ! in user messages.
Shorted a number of self.caller.msg to caller.msg in commands.
If a player tries to wear an object not in person but in room, they receive a suggestion to get the item.
Removed CmdGive. Players should not have a method to forcibly give another an object. The evennia give will later be removed.