draft standard command set
created outline of a command set for standard commands. put inventory and drop commands in it, from the clothing module.
Moved unit tests for clothing module to the commands unit test. As well as the StandaardCmdsCmdSet commands drop and inventory. Likely both commands will have futher tests as other object types are created.