commad not target help & wear cmd update
Command module will show a help message if a target is needed and none was provided.
Removed TestClothingCmd and put clothing commands tests in the general TestCommands module
wear command:
removed support for styles. Mostly to avoid the rude text messages the general public would use.
removed the need to collect a target for the command