Command bugfix
commands.command.unarmed, did not have a proper format for dmg_types. was dmg_types = ("BLG"), needed to be dmg_types = ("BLG",). Without the , it just returns a string.

commands.command.combat_command, was converting part_hit from a list to a string too late.

world.rules.damage.get_dmg_after_dr, was not verifing a body part existed when checking for body part dr of a certain type. It should have been erroring when attacking objects. Somehow it was not.
Also get_dmg_after_dr will now throw a warning message when a body part is not found on a body instance when it was in that targets body.parts list.

unit tests pass in 10.605s and 10.564s