rpsystem pose, sdesc, mask permision blocked
created
an overridden command set for the rp system.
Did so to lock out the pose sdesc and mask commands. They are not builder commands.
Overrode pose sdesc and mask to make them builder or higher permission

created unit local test for rp commands. This is in additional to the box tests previously copied over.
Make unit test to verify these command permisions

unit tests pass in 10.907s, and 10.927s