Unconscious Character pose
Unconscious Character pose that they are unconscious. Will show when another Character looks while in a room.
A message shows to player and everyone in room when Character falls into or recovers from unconsciousness.

Added unit test for pose. Had to be in commands.tests could not get it to work in typeclasses unit test.

unit tests pass in 11.170s, and 11.107s