body.get_part returns an instance
body.get_part returns an instance of the part.
Allows for rest of a specific body part.

rules.damage.get_dmg_after_dr, was changed to accept an instance and string for the body_part argument.
Also has an argument for target, to support commands that target multiple targets.

typeclasses.mixins was changed to support the update to body.get_part

Changed to unit tests to support update.

Unit tests pass in 13.997s and 14.756s