holding and CmdWear bugfix
Character.isHolding returns an instance of the hand holding the object passed. Update from returning True if the object is held.
Character.hands(), If a held object was removed from the database, it is now removed an unwielding from hands automatically.

clothing.wear returns True if the clothing was successfully worn.
CmdWear will empty (or unoccupy) the hand an article of successfully worn clothing was in.

Added unit tests for:
auto unholding and wielding a delete item
Removing an item from hand and wielding after worn.

unit tests pass in 15.204s and 15.562s